click on a row

+++ app/src/main/java/com/example/cookingntraveler/api/FilterRecipe.kt
@@ -6,5 +6,5 @@ import com.google.gson.annotations.SerializedName
 data class FilterRecipe (
     @SerializedName("strMeal") val strMeal : String,
     @SerializedName("strMealThumb") val strMealThumb : String,
-    @SerializedName("idMeal")val idMeal : Int
+    @SerializedName("idMeal")val idMeal : Long
 )
